Fix RoutePlayer location update race
RoutePlayer was kicking off a new Job to monitor the location each time
a route was started. This was resulting in multiple coroutines executing
on the same data on every location update - obviously very bad. This
change has fixed that, with the additional benefit that the RoutePlayer
location monitoring now only starts and stops when a route starts and
stops.