The proximity beacon needs to run at the same time as the heading beacon
I had mis-understood how the proximity beacon was supposed to work, so this
addresses that issue. Beacons created from Kotlin can now consist of two
separated beacons - one for the heading sound and one for a proximity
sound.
The proximity beacon is only created when playing back a route, and most of
the time it is playing silence. However, once the user gets close it starts
playing its little hum, changing again as it gets closer.
There's also changes to handle beacons that are stereo audio (like the
proximity tones).