Fix PlacesNearby to use the point nearest to the users location as the marker
This didn't affect POI which are points e.g. bus stops because there is only
one point that can be chosen. However, POI which are lines or polygons were
not always having the correct point chosen - that which is nearest to the user.
This change fixes that.

It can be argued that using a centroid or an Entrance if there is one would
make more sense, but that's another issue.