FMOD audio attentuation settings were all wrong
These changes don't affect the audio (much), but they do at least mean
that the FMOD settings now match the coordinate system being used. We
used to use raw GPS coordinates, but that resulted in incorrect headings
and so we now use x and y coordinates in metres. The changes mean that
in theory sounds could get softer if they are positioned more than 10m
and less than 500m away. However, the attenuation per metre is set to
0.01 and so it currently has very little effect. However, we could
easily add a setting to control this if users want far away beacons to
be softer.