More finessing of offline map downloads
A .progress file is created so that the app knows when a download is taking place.
If Soundscape is exited then the download carries on going. When Soundscape starts
up again it checks for any .progress files and if there is one it goes back to the
offline maps downloading screen. The aim is to keep the user aware of the downloading
file.