Add dynamic map source selection to MapContainerLibre
Prior to this change, the `style.json` files were being pre-processed with the Protomaps
sources added in. However, it's much easier to do this dynamically within MapContainerLibre.
When the style is loaded we now add the vectorSource directly. This means that we should be
able to easily open maps from different sources and that should make switching between
extracts trivial.