Improve entrance matching behaviour
Several improvements here:

1. Try and match the `layer` of buildings and entrances. This disambiguates
some entrances where there are multiple layers of POIs that it coincides with.
2. Don't include entrances with no name and no way of calculating one.
3. Don't include entrances where there are multiple possibilities as to what it
should be named e.g. the building has more than one name. It's not possible to
get this correct, and it's better not to be wrong. The fix in OSM data is to
name the entrance.