Duplicate Features to have one for each Entrance
When finding entrances the code now creates a copy of the POI feature for
each entrance and marks the original POI feature as containing entrances.
This means that PlacesNearby can now filter out any POI which has entrances
and display the entrances instead which is more useful.
Because this only affects POIs with entrances and not named buildings, the
number of entrances in the tests has dropped again.