Remove crashing Log message in StorageUtils
Confusingly in Kotlin, File can be null despite not being File?. The crash
almost certainly is happening on a phone which has an empty SD slot. In that
case File is null and we were trying to use it, catching the exception and
then trying to use it again in the Log message - doh!
I've fixed the Log message, but also added code to skip even trying null
Files.