Move to using Retrofit for offline map downloads
The Android Download Manager isn't configurable for timeouts and retries,
and this was causing problems when a large extract was being downloaded for
the first time. The code has now moved to use Retrofit and if the server
returns 503 responses it will indicate to the user that the server is
getting ready to return data. It will retry multiple times after an initial
delay so that it should manage to start the download successfully.
The initial connect timeout has also been set to 3 minutes so that it
should work with the current server which isn't returning 503s.