Trim data and time from start of filenames used to save extracts
The server provides extracts with names like:

    20251106-1128-glasgow-gb.pmtiles

This change strips the date and time from the start. This means that we
don't risk having two extracts for the same region from different map
builds. We also don't delete the previous extract until we have
successfully downloaded the new extract. This does mean that twice the
storage is required, but it should help to avoid accidentally blowing
away downloaded extracts.