STA-227 WIP attempt to do structure-as-text regression tests.
Tried this to copy to emulated SD card, but that doesn't work on recent Android SDKs.

ChatGPT's next suggestion was to gather all the results to a TAR file on the Android side and stream it out via adb, but I think I can just get the test to write diffs to the Gradle build folder directly, instead of to the Android filesystem.