Tidy up TTS engine changing
The TextToSpeech initialization code had got very confusing. One of the main
confounding factors was that the LanguageViewModel from onboarding was sneaked
into Settings and it was creating a second audio engine. It's great that the
code can handle that, but less confusing when you're trying to understand why
the logging for TTS initialisation is all duplicated...
The changes here are:
* Remove the audio engine from LanguageViewModel which wasn't in fact doing
anything anyway. The audio engine only lived for the life of the language
selection and was never actually used.
* Remove all of the shared preference watching in TtsEngine which was simply
duplicating that in NativeAudioEngine.
* Improve the NativeAudioEngine behaviour so that TTS initialisation happens
in a more predictable way.

I'm hoping that this will address a user issue where he couldn't select an
engine on his high spec Samsung Android 16 phone, but at least the code is
much clearer now.