More backup tweaking
Because we target Android 11 (sdk 30) the backup apparently only uses
backup_rules.xml so I've removed the Android 12 data_extraction_rules
for now. The backup_rules.xml are tested to work correctly with Android
16 and Android 11 on an emulated Pixel. However, I can't get the restore
to work on a Samsung device, though it does seem to at least backup the
correct pieces of data. I'm going to assume that this is a an issue with
the test_local_backup.sh approach on Samsung devices and won't affect
real life backups.