Text entry edit box semantics override broke advanced editing features
This is a good case of Android knows best. I was trying to improve the text
edit hints to be more useful, but in the process I managed to disable some
of the more advanced features such as copy and paste. I think the best
approach for now is to simply leave it at the defaults. Verbosity of hints
can be controlled by the user and my Talkback using experience is too limited
to be able to sensibly adjust this.