Add zipped logfile to Contact Support request
We haven't had enough visibility into some app issues, so the aim here is
to provided as much logging as we can. The logs can be fairly large so we:

1. Request that logcat only gives us the last 5000 lines of log
1. Filter out all but the last 30 minutes of logging
2. Zip the log

The log is formatted as JSON that can be directly imported in Android Studio logcat
pane. The exports directory is deleted with all its contents whenever the app starts
up. This prevents an accumulation of exported GPX and log files.