Keep track of untranslated OSM tag values and add to Contact Support
This change adds OSM Strings for which we have no translation to a Set which
is returned if the user uses Contact Support. This should be a relatively easy
way to grow our translation to cover real OSM values which are out there in the
wild. Note that the values are only used when a POI doesn't have a name, so for
things like shops this may be relatively rare. However, if we add the type of
shop to the callouts e.g. "Spar convenience store" then it will become more
common.