Add more flexible geocoding including using offline tiles
Tests for both forward and reverse geocoding including use of a library
(AndroidAddressFormatter) to present the addresses in the correct format
for their location. That is now used for Photon, Android and Offline output
and the results tend to be fairly close.

During tile parsing we now parse the housenumber layer into a number of
FeatureTrees, one for each named road. This makes it much easier to reverse
geocode when we're map matched - or indeed just assuming the nearest road.
The Geocoder Test now gives the most accurate results with LocalDecoder for
the first test where the location is on a corner between two streets. When
there's a map matched Way the LocalDecoder uses it, whereas the other two
are performing a search without any extra knowledge.

There's also a new StreetDescription class. This follows a named street and
creates linear maps of street numbers and point of interest. This can then
be looked up to get an interpolated street address and a description of the
location relative to intersections and POI. It's used within OfflineGeocoder
for reverse geocoding.

The search UI has been changed so that the user has to hit the search icon
to trigger a search. There are a few reasons for this:

 1. AndroidGeocoder calls have a daily allowed limit and so we don't want to
  use more than we have to.
 2. It encourages the user to write a more complete search which is what is
  required by both AndroidGeocoder and LocalGeocoder.

The default geocoder setup is "Auto" which uses Android if there's network,
and Offline if there's not i.e. not Photon at all. However, in theory not all
phones have Android geocoder support and so it will fall back to using Photon
if that's the case and there is network. Settings has the option to select
which search to use. Offline will still be the fall back in the case of no
network. Offline search results have a slightly different icon in the results
list (just for developer debug for now).

The string comparison used in searching is an adjusted Levenshtein Damerau which
compares only up to the length of the search string. This is to allow matching
of partial strings e.g. "Tesco" (or indeed "Yesco" with "Tesco express").